refactor(utils): split in mutliple files
